TBW-74 fixed issues in language menu and small key fixes
