+++ frontend/src/Components/CustomToolbar/CustomToolbar.jsx
@@ -3,7 +3,8 @@ import { Link } from "react-router-dom";
 import { AppBar, Menu, MenuItem, Toolbar, Button, Tabs, Tab } from "@material-ui/core";
 import { useLocation } from "react-router-dom";
 import { useTranslation } from "react-i18next";
-import { returnFlagByLanguage, RouterPaths } from "Utils/";
+import { RouterPaths } from "Utils/";
+import { LanguageMenu } from "Components/";
 import MenuIcon from "@material-ui/icons/Menu";
 import MediaQuery from "react-responsive";
 import "./CustomToolbar.scss";
@@ -11,9 +12,8 @@ import logo from "Assets/images/tbw-logo.png";
 
 const CustomToolbar = props => {
   const location = useLocation();
-  const { t, i18n } = useTranslation();
+  const { t } = useTranslation();
   const [toolbarColor, setToolbarColor] = useState("#0b121000");
-  const [anchorLng, setAnchorLng] = React.useState(null);
   const [anchorMn, setAnchorMn] = React.useState(null);
   const [currentPage, setCurrentPage] = React.useState(location.pathname.replace("/", ""));
   const { notLanding } = props;
@@ -34,15 +34,6 @@ const CustomToolbar = props => {
 
   const handleChange = (event, newValue) => {
     setCurrentPage(newValue);
-  };
-
-  const handleLngClick = event => {
-    setAnchorLng(event.currentTarget);
-  };
-
-  const handleLngClose = lang => {
-    i18n.changeLanguage(lang);
-    setAnchorLng(null);
   };
 
   const handleMnClick = event => {
@@ -82,6 +73,7 @@ const CustomToolbar = props => {
             {options.map(value => {
               return (
                 <Tab
+                  key={value}
                   value={value.toLowerCase()}
                   label={value}
                   to={`/${value.toLowerCase()}`}
@@ -91,31 +83,7 @@ const CustomToolbar = props => {
             })}
           </Tabs>
         </MediaQuery>
-        <Button
-          className="language-button"
-          aria-controls="lng-menu"
-          aria-haspopup="true"
-          onClick={handleLngClick}
-        >
-          {returnFlagByLanguage(i18n.language)}
-          {i18n.language}
-        </Button>
-        <Menu
-          id="lng-menu"
-          anchorEl={anchorLng}
-          keepMounted
-          open={Boolean(anchorLng)}
-          onClose={handleLngClose}
-        >
-          <MenuItem onClick={() => handleLngClose("en")}>
-            {returnFlagByLanguage("en")}
-            EN
-          </MenuItem>
-          <MenuItem onClick={() => handleLngClose("fr")}>
-            {returnFlagByLanguage("fr")}
-            FR
-          </MenuItem>
-        </Menu>
+        <LanguageMenu />
 
         <MediaQuery query="(max-width: 1023px)">
           <Button
