+++ frontend/src/Components/CustomToolbar/CustomToolbar.scss
@@ -36,7 +36,7 @@
     font-weight: 300 !important;
   }
 
-  .flagIcon {
+  .flag-icon {
     margin-right: 2px;
     height: 12px;
   }
