+++ frontend/src/Components/index.js
@@ -7,3 +7,4 @@ export { default as CustomServicesNav } from "./CustomServicesNav";
 export { default as Project } from "./Project";
 export { default as ContactForm } from "./ContactForm";
 export { default as CareersForm } from "./CareersForm";
+export { default as LanguageMenu } from "./LanguageMenu";
